+++ src/string.ts
@@ -1,0 +1,14 @@
+import { IDate, ITime, minutesInHour } from "./datetime";
+
+const pad0 = (n: number, len: number) => String(n | 0).padStart(len, "0");
+
+export const dateToString = (date: IDate): string => {
+    return pad0(date.year, 4) + "-" + pad0(date.month, 2) + "-" + pad0(date.day, 2);
+};
+
+export const timeToString = (time: ITime): string => {
+    return pad0(time.hour, 2)
+        + ":" + pad0(time.minute, 2)
+        + ":" + pad0(time.second, 2)
+        + "." + pad0(time.millisecond, 3);
+};